add an explicit background to select
The user agent may define his own background by default and also his own foreground color for the dropdown. If the user-agent does not allow to overwrite background and color that's fine. But at least chrome inherits by default the foreground color from the select but not the background color.

In this case it would mean white text on a white background which is not that good for reading.